Create redirection headers
